Refactor openssl integration
