+++ tlspuffin/src/openssl/deterministic.rs
@@ -1,0 +1,24 @@
+use log::warn;
+use std::os::raw::c_int;
+
+#[cfg(feature = "openssl111")]
+extern "C" {
+    fn make_openssl_deterministic();
+    fn RAND_seed(buf: *mut u8, num: c_int);
+}
+
+#[cfg(feature = "openssl111")]
+pub fn set_openssl_deterministic() {
+    warn!("OpenSSL is no longer random!");
+    unsafe {
+        make_openssl_deterministic();
+        let mut seed: [u8; 4] = 42u32.to_le().to_ne_bytes();
+        let buf = seed.as_mut_ptr();
+        RAND_seed(buf, 4);
+    }
+}
+
+#[cfg(not(feature = "openssl111"))]
+pub fn make_openssl_deterministic() {
+    warn!("Failed to make PUT determinisitic!");
+}
