+++ tlspuffin/src/openssl/mod.rs
@@ -1,19 +1,34 @@
+use std::io::ErrorKind;
 use std::{cell::RefCell, io, rc::Rc};
 
-use openssl::error::ErrorStack;
+use openssl::{
+    error::ErrorStack,
+    pkey::{PKeyRef, Private},
+    ssl::{Ssl, SslContext, SslMethod, SslStream},
+    x509::X509Ref,
+};
 use rustls::msgs::message::OpaqueMessage;
 
+use crate::openssl::deterministic::set_openssl_deterministic;
+use crate::openssl::util::{set_max_protocol_version, static_rsa_cert};
 use crate::{
-    agent::{AgentName, PutName},
-    concretize::{Config, Put},
+    agent::{AgentName, PutName, TLSVersion},
     error::Error,
     io::{MemoryStream, MessageResult, Stream},
-    openssl::openssl_binding::{openssl_version, static_rsa_cert},
-    registry::{Factory, OPENSSL111},
+    put::{Config, Put},
+    put_registry::{Factory, OPENSSL111},
     trace::VecClaimer,
 };
 
-mod openssl_binding;
+mod deterministic;
+mod util;
+
+/*
+   Change openssl version:
+   cargo clean -p openssl-src
+   cd openssl-src/openssl
+   git checkout OpenSSL_1_1_1j
+*/
 
 pub fn new_openssl_factory() -> Box<dyn Factory> {
     struct OpenSSLFactory;
@@ -38,14 +53,14 @@ pub fn new_openssl_factory() -> Box<dyn Factory> {
     Box::new(OpenSSLFactory)
 }
 
-impl From<openssl::error::ErrorStack> for Error {
+impl From<ErrorStack> for Error {
     fn from(err: ErrorStack) -> Self {
         Error::OpenSSL(err.to_string())
     }
 }
 
 pub struct OpenSSL {
-    stream: openssl::ssl::SslStream<MemoryStream>,
+    stream: SslStream<MemoryStream>,
 }
 
 impl Drop for OpenSSL {
@@ -66,36 +81,33 @@ impl Stream for OpenSSL {
 }
 
 impl io::Read for OpenSSL {
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
+    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
         self.stream.get_mut().read(buf)
     }
 }
 
 impl io::Write for OpenSSL {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
+    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
         self.stream.get_mut().write(buf)
     }
 
-    fn flush(&mut self) -> std::io::Result<()> {
+    fn flush(&mut self) -> io::Result<()> {
         self.stream.get_mut().flush()
     }
 }
 
 impl Put for OpenSSL {
     fn new(config: Config) -> Result<OpenSSL, Error> {
-        let memory_stream = MemoryStream::new();
-        let stream = if config.server {
+        let ssl = if config.server {
             //let (cert, pkey) = openssl_binding::generate_cert();
             let (cert, pkey) = static_rsa_cert()?;
-            openssl_binding::create_openssl_server(
-                memory_stream,
-                &cert,
-                &pkey,
-                &config.tls_version,
-            )?
+
+            Self::create_server(&cert, &pkey, config.tls_version)?
         } else {
-            openssl_binding::create_openssl_client(memory_stream, &config.tls_version)?
+            Self::create_client(config.tls_version)?
         };
+
+        let stream = SslStream::new(ssl, MemoryStream::new())?;
 
         #[cfg(not(feature = "claims"))]
         let openssl = OpenSSL { stream };
@@ -110,7 +122,15 @@ impl Put for OpenSSL {
     }
 
     fn progress(&mut self) -> Result<(), Error> {
-        openssl_binding::do_handshake(&mut self.stream)
+        if self.is_state_successful() {
+            // Trigger another read
+            let mut vec: Vec<u8> = Vec::from([1; 128]);
+            let maybe_error: MaybeError = self.stream.ssl_read(&mut vec).into();
+            maybe_error.into()
+        } else {
+            let maybe_error: MaybeError = self.stream.do_handshake().into();
+            maybe_error.into()
+        }
     }
 
     fn reset(&mut self) -> Result<(), Error> {
@@ -163,10 +183,117 @@ impl Put for OpenSSL {
     }
 
     fn version() -> &'static str {
-        openssl_version()
+        openssl::version::version()
     }
 
     fn make_deterministic() {
-        openssl_binding::make_deterministic();
+        set_openssl_deterministic();
+    }
+}
+
+impl OpenSSL {
+    fn create_server(
+        cert: &X509Ref,
+        key: &PKeyRef<Private>,
+        tls_version: TLSVersion,
+    ) -> Result<Ssl, ErrorStack> {
+        let mut ctx_builder = SslContext::builder(SslMethod::tls())?;
+        ctx_builder.set_certificate(cert)?;
+        ctx_builder.set_private_key(key)?;
+
+        #[cfg(feature = "openssl111")]
+        ctx_builder.clear_options(openssl::ssl::SslOptions::ENABLE_MIDDLEBOX_COMPAT);
+
+        #[cfg(feature = "openssl111")]
+        ctx_builder.set_options(openssl::ssl::SslOptions::ALLOW_NO_DHE_KEX);
+
+        set_max_protocol_version(&mut ctx_builder, tls_version)?;
+
+        #[cfg(any(feature = "openssl101f", feature = "openssl102u"))]
+        {
+            ctx_builder.set_tmp_ecdh(
+                openssl::ec::EcKey::from_curve_name(openssl::nid::Nid::SECP384R1)
+                    .as_ref()
+                    .unwrap(),
+            )?;
+            // TODO: https://github.com/sfackler/rust-openssl/issues/1529 use callback after fix
+            //ctx_builder.set_tmp_ecdh_callback(|_, _, _| {
+            //   openssl::ec::EcKey::from_curve_name(openssl::nid::Nid::SECP384R1)
+            //});
+        }
+
+        #[cfg(any(feature = "openssl101f", feature = "openssl102u"))]
+        {
+            ctx_builder.set_tmp_rsa(openssl::rsa::Rsa::generate(512).as_ref().unwrap())?;
+            // TODO: https://github.com/sfackler/rust-openssl/issues/1529 use callback use callback after fix
+            //ctx_builder.set_tmp_rsa_callback(|_, is_export, keylength| openssl::rsa::Rsa::generate(keylength));
+        }
+
+        // Allow EXPORT in server
+        ctx_builder.set_cipher_list("ALL:EXPORT:!LOW:!aNULL:!eNULL:!SSLv2")?;
+
+        let mut ssl = Ssl::new(&ctx_builder.build())?;
+        ssl.set_accept_state();
+
+        Ok(ssl)
+    }
+
+    fn create_client(tls_version: TLSVersion) -> Result<Ssl, ErrorStack> {
+        let mut ctx_builder = SslContext::builder(SslMethod::tls())?;
+        // Not sure whether we want this disabled or enabled: https://gitlab.inria.fr/mammann/tlspuffin/-/issues/26
+        // The tests become simpler if disabled to maybe that's what we want. Lets leave it default
+        // for now.
+        // https://wiki.openssl.org/index.php/TLS1.3#Middlebox_Compatibility_Mode
+        #[cfg(feature = "openssl111")]
+        ctx_builder.clear_options(openssl::ssl::SslOptions::ENABLE_MIDDLEBOX_COMPAT);
+
+        set_max_protocol_version(&mut ctx_builder, tls_version)?;
+
+        // Disallow EXPORT in client
+        ctx_builder.set_cipher_list("ALL:!EXPORT:!LOW:!aNULL:!eNULL:!SSLv2")?;
+
+        let mut ssl = Ssl::new(&ctx_builder.build())?;
+        ssl.set_connect_state();
+
+        Ok(ssl)
+    }
+}
+
+pub enum MaybeError {
+    Ok,
+    Err(Error),
+}
+
+impl<T> From<Result<T, openssl::ssl::Error>> for MaybeError {
+    fn from(result: Result<T, openssl::ssl::Error>) -> Self {
+        if let Err(ssl_error) = result {
+            if let Some(io_error) = ssl_error.io_error() {
+                match io_error.kind() {
+                    ErrorKind::WouldBlock => {
+                        // Not actually an error, we just reached the end of the stream, thrown in MemoryStream
+                        // trace!("Would have blocked but the underlying stream is non-blocking!");
+                        MaybeError::Ok
+                    }
+                    _ => MaybeError::Err(Error::IO(format!("Unexpected IO Error: {}", io_error))),
+                }
+            } else if let Some(ssl_error) = ssl_error.ssl_error() {
+                // OpenSSL threw an error, that means that there should be an Alert message in the
+                // outbound channel
+                MaybeError::Err(Error::OpenSSL(ssl_error.to_string()))
+            } else {
+                MaybeError::Ok
+            }
+        } else {
+            MaybeError::Ok
+        }
+    }
+}
+
+impl Into<Result<(), Error>> for MaybeError {
+    fn into(self) -> Result<(), Error> {
+        match self {
+            MaybeError::Ok => Ok(()),
+            MaybeError::Err(err) => Err(err),
+        }
     }
 }
